Move state.hpp into separate directory for better organizing
