+++ example/quantum_phase_estimation_diagonal.cpp
@@ -12,7 +12,7 @@
 #include <mini-qiskit/gates/fourier.hpp>
 #include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
-#include <mini-qiskit/state.hpp>
+#include <mini-qiskit/state/state.hpp>
 
 /*
     We first create a 4x4 diagonal unitary matrix, where the diagonal elements are
