+++ example/quantum_phase_estimation_trotter20.cpp
@@ -13,7 +13,7 @@
 #include <mini-qiskit/gates/fourier.hpp>
 #include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
-#include <mini-qiskit/state.hpp>
+#include <mini-qiskit/state/state.hpp>
 #include <mini-qiskit/io/statevector.hpp>
 
 /*
