+++ example/quantum_teleportation.cpp
@@ -3,7 +3,7 @@
 #include <random>
 
 #include <mini-qiskit/circuit/circuit.hpp>
-#include <mini-qiskit/state.hpp>
+#include <mini-qiskit/state/state.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
 #include <mini-qiskit/calculations/measurements.hpp>
