+++ example/read_decomposition_file.cpp
@@ -8,7 +8,7 @@
 #include <mini-qiskit/common/print.hpp>
 #include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
-#include <mini-qiskit/state.hpp>
+#include <mini-qiskit/state/state.hpp>
 
 auto main() -> int
 {
