+++ include/mini-qiskit/common/print.hpp
@@ -13,7 +13,7 @@
 #include "mini-qiskit/primitive_gate.hpp"
 #include "mini-qiskit/common/matrix2x2.hpp"
 #include "mini-qiskit/circuit/circuit.hpp"
-#include "mini-qiskit/state.hpp"
+#include "mini-qiskit/state/state.hpp"
 
 
 namespace impl_mqis
