+++ include/mini-qiskit/io/numpy_statevector.hpp
@@ -5,7 +5,7 @@
 #include <iostream>
 #include <vector>
 
-#include <mini-qiskit/state.hpp>
+#include <mini-qiskit/state/state.hpp>
 
 namespace impl_mqis
 {
