+++ include/mini-qiskit/io/statevector.hpp
@@ -6,7 +6,7 @@
 #include <string>
 #include <vector>
 
-#include <mini-qiskit/state.hpp>
+#include <mini-qiskit/state/state.hpp>
 
 namespace impl_mqis
 {
