+++ include/mini-qiskit/mini-qiskit.hpp
@@ -5,4 +5,4 @@
 #include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/primitive_gate.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
-#include <mini-qiskit/state.hpp>
+#include <mini-qiskit/state/state.hpp>
