+++ qpe_tangelo_project/qpe_tangelo_simulate.cpp
@@ -5,7 +5,7 @@
 #include <mini-qiskit/circuit_operations/append_circuits.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
 #include <mini-qiskit/circuit/circuit.hpp>
-#include <mini-qiskit/state.hpp>
+#include <mini-qiskit/state/state.hpp>
 #include <mini-qiskit/io/statevector.hpp>
 
 /*
