+++ test/source/calculations/probabilities_test.cpp
@@ -12,7 +12,7 @@
 #include "mini-qiskit/calculations/probabilities.hpp"
 #include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/simulation/simulate.hpp"
-#include "mini-qiskit/state.hpp"
+#include "mini-qiskit/state/state.hpp"
 
 static constexpr auto RELATIVE_TOL = 1.0e-6;
