+++ test/source/circuit_operations/make_controlled_circuit_test.cpp
@@ -2,7 +2,7 @@
 #include <catch2/generators/catch_generators.hpp>
 
 #include "mini-qiskit/circuit/circuit.hpp"
-#include "mini-qiskit/state.hpp"
+#include "mini-qiskit/state/state.hpp"
 #include "mini-qiskit/simulation/simulate.hpp"
 #include "mini-qiskit/gates/common_u_gates.hpp"
 #include "mini-qiskit/gates/multiplicity_controlled_u_gate.hpp"
