+++ test/source/decomposed/build_decomposed_circuit_test.cpp
@@ -13,7 +13,7 @@
 #include <mini-qiskit/decomposed/build_decomposed_circuit.hpp>
 #include <mini-qiskit/circuit/circuit.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
-#include <mini-qiskit/state.hpp>
+#include <mini-qiskit/state/state.hpp>
 
 
 template <typename Container, typename Lambda>
