+++ test/source/gates/fourier_test.cpp
@@ -10,7 +10,7 @@
 #include "mini-qiskit/common/mathtools.hpp"
 #include "mini-qiskit/gates/fourier.hpp"
 #include "mini-qiskit/simulation/simulate.hpp"
-#include "mini-qiskit/state.hpp"
+#include "mini-qiskit/state/state.hpp"
 
 TEST_CASE("Forward QFT on |0> state")
 {
