+++ test/source/simulation/measure_test.cpp
@@ -10,7 +10,7 @@
 
 #include <mini-qiskit/common/print.hpp>
 #include <mini-qiskit/circuit/circuit.hpp>
-#include <mini-qiskit/state.hpp>
+#include <mini-qiskit/state/state.hpp>
 #include <mini-qiskit/simulation/simulate.hpp>
 #include <mini-qiskit/simulation/measure.hpp>
