+++ test/source/state_test.cpp
@@ -11,7 +11,7 @@
 
 #include "mini-qiskit/circuit/circuit.hpp"
 #include "mini-qiskit/simulation/simulate.hpp"
-#include "mini-qiskit/state.hpp"
+#include "mini-qiskit/state/state.hpp"
 
 TEST_CASE("QuantumState endian representation")
 {
